Update and rename README.md to ATP_LPP_en.py
English version of the problem solution. You can enter the following data: supplier Capacities, customer requests, and a table of transportation prices. The algorithm determines the type of problem and, if necessary, reduces it to a linear programming problem, then passes the Packed data to the linprog module of the SciPy package.